ADD. [MAIN] - Hardware Abstract 데몬 상태전이에 연동

+++ daemon/hardware/abstract/knob.h
@@ -9,5 +9,5 @@
 
 class knob_abstract {
 protected: /* abstract */
-	virtual uint8_t read() = 0;
+	virtual uint16_t read() = 0;
 };
